requirements.txt
# Horme POV Production Requirements
# Enterprise AI Recommendation System
# ==============================================================================

# Core Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database - PostgreSQL
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Database - Neo4j Knowledge Graph
neo4j==5.16.0

# Cache - Redis
redis==5.0.1
hiredis==2.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.0
pyjwt==2.8.0

# OpenAI LLM Integration
openai==1.51.2

# NLP & ML Libraries
sentence-transformers==2.3.1
transformers==4.37.0
torch==2.1.2
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.1.4

# Vector Database & Embeddings
chromadb==0.4.22
langchain==0.1.0
langchain-openai==0.0.5

# Data Processing
openpyxl==3.1.2  # Excel file support
xlsxwriter==3.1.9
python-dateutil==2.8.2

# Document Processing & PDF Generation
pdfplumber==0.11.0  # PDF text extraction
python-docx==1.1.0  # DOCX file support
reportlab==4.0.9  # PDF generation

# HTTP & API Clients
httpx==0.26.0
requests==2.31.0
aiohttp==3.9.1

# Logging & Monitoring
structlog==24.1.0
python-json-logger==2.0.7
prometheus-client==0.19.0

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# WebSocket Support
websockets==12.0
python-socketio==5.11.0

# Background Tasks
celery==5.3.6
# ✅ PRODUCTION FIX: Removed duplicate redis (already listed on line 22)

# Testing (Development)
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
# ✅ PRODUCTION FIX: Removed duplicate httpx (already listed on line 53)
faker==22.5.1

# Code Quality (Development)
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.6

# Utilities
tenacity==8.2.3  # Retry logic
cachetools==5.3.2  # Caching utilities
python-slugify==8.0.1
pytz==2024.1
ujson==5.9.0  # Fast JSON
orjson==3.9.12  # Faster JSON

# Production WSGI Server
gunicorn==21.2.0

# Environment & Process Management
supervisor==4.2.5

# Data Validation
email-validator==2.1.0
phonenumbers==8.13.27

# Rate Limiting
slowapi==0.1.9

# CORS
fastapi-cors==0.0.6

# Health Checks
py-healthcheck==1.10.1

# Metrics
statsd==4.0.1

# AWS SDK (if deploying to AWS)
# boto3==1.34.25
# botocore==1.34.25

# ==============================================================================
# NOTES:
# ==============================================================================
# 1. Install with: pip install -r requirements.txt
# 2. For development, also install dev dependencies
# 3. Use virtual environment: python -m venv venv && source venv/bin/activate
# 4. Freeze current versions: pip freeze > requirements.lock
# 5. Update all: pip install --upgrade -r requirements.txt
#
# Docker Usage:
# - This file is used in Dockerfile.api
# - Ensure all dependencies are compatible with Python 3.11
# - Test in Docker container before deploying to production
#
# Performance Considerations:
# - sentence-transformers includes large models (~80MB-400MB)
# - torch adds ~1GB to image size
# - Consider multi-stage Docker builds to minimize final image size
# ==============================================================================
